Shift common code from home into app and Add widget files
