+++ frontend/src/styles/style.scss
@@ -50,6 +50,10 @@ a.broken-link {
   padding: 2rem 1.5rem;
 }
 
+.widget-padding {
+  padding: .5rem 1rem;
+}
+
 .warn {
   color: mui-color('red');
