+++ frontend/src/styles/summary-chart.scss
@@ -1,0 +1,225 @@
+@import 'colors';
+
+/* Summary */
+#summary {
+  .summary-status {
+    text-align: center;
+  }
+
+  @mixin icon-button-config {
+    color: mui-color("grey");
+    padding: 0 1.2px 0 1.2px;
+    text-decoration: none;
+  }
+
+  .icon-button {
+    @include icon-button-config;
+    cursor: pointer;
+  }
+
+  .broken-link {
+    .icon-button {
+      cursor: default;
+    }
+  }
+
+  .summary-picker {
+    align-items: center;
+    display: flex;
+    flex-flow: row wrap;
+    justify-content: center;
+    margin-bottom: 2rem;
+
+    &__section {
+      align-items: inherit;
+      display: flex;
+      flex: 0 1 auto;
+      flex-flow: inherit;
+      justify-content: inherit;
+    }
+
+    &__checkboxes {
+      label {
+        margin-left: 0.5rem;
+      }
+
+      span {
+        margin-left: 0.25rem;
+      }
+    }
+
+    .mui-textfield,
+    .mui-select {
+      margin: 0.5rem;
+      padding-right: 10px;
+    }
+
+    .mui-btn {
+      background: transparent;
+      box-shadow: none;
+      color: mui-color("grey");
+      font-size: 0.75rem;
+      font-weight: bold;
+      left: -8px;
+      margin: 0;
+      padding: 0;
+      vertical-align: middle;
+    }
+
+    .search_box {
+      align-items: center;
+      display: flex;
+    }
+
+    input {
+      font-size: 0.75rem;
+      padding-right: 10px;
+    }
+
+    label {
+      font-size: 0.75rem;
+      overflow-y: hidden;
+      text-align: left;
+      width: fit-content;
+    }
+
+    input,
+    select {
+      font-size: 0.75rem;
+    }
+  }
+
+  .summary-charts {
+    margin-bottom: 1.4rem;
+
+    &__title {
+      align-items: center;
+      display: flex;
+      font-weight: bold;
+      text-align: left;
+
+      & > * {
+        padding-right: 0.5rem;
+      }
+
+      &--index {
+        background: mui-color("black");
+        color: mui-color("white");
+        font-size: medium;
+        overflow: hidden;
+        padding: 0.1em 0.25em;
+        vertical-align: middle;
+      }
+
+      &--groupname {
+        font-size: medium;
+        padding: 0.5rem;
+      }
+
+      &--percentile {
+        @include mini-font;
+        color: mui-color("grey");
+        margin-left: auto;
+      }
+
+      &--contribution {
+        @include mini-font;
+        display: inline;
+      }
+    }
+
+    &__fileType--breakdown {
+      overflow-y: hidden;
+
+      &__legend {
+        display: inline;
+        float: left;
+      }
+    }
+  }
+
+  .summary-chart {
+    display: inline-block;
+    margin-bottom: 1rem;
+    position: relative;
+    text-align: left;
+    width: 100%;
+
+    &__title {
+      align-items: center;
+      clear: left;
+      display: flex;
+
+      & > * {
+        padding-right: 0.5rem;
+      }
+
+      &--index {
+        margin-left: 3px;
+      }
+
+      &--repo {
+        font-weight: bold;
+      }
+
+      &--index::after {
+        content: ".";
+      }
+
+      &--repo {
+        padding-right: 0.25rem;
+      }
+
+      &--contribution {
+        @include mini-font;
+      }
+
+      &--percentile {
+        @include mini-font;
+        color: mui-color("grey");
+        margin-left: auto;
+        padding-right: 0;
+      }
+    }
+
+    &__ramp {
+      position: relative;
+
+      .overlay {
+        height: 100%;
+        position: absolute;
+        top: 0;
+
+        &.show {
+          background-color: rgba(mui-color("white"), 0.5);
+          border: 1px dashed mui-color("black");
+        }
+
+        &.edge {
+          border-right: 1px dashed mui-color("black");
+        }
+      }
+    }
+
+    &__contrib {
+      text-align: left;
+
+      &--bar {
+        background-color: mui-color("red");
+        float: left;
+        height: 4px;
+        margin-top: 2px;
+      }
+    }
+  }
+
+  .active-icon {
+    background-color: mui-color("green");
+    border-radius: 2px;
+    color: mui-color("white");
+  }
+
+  .active-background {
+    background-color: mui-color("yellow", "200");
+  }
+}
