+++ frontend/src/views/c-widget.vue
@@ -1,0 +1,66 @@
+<template lang="pug">
+template(v-if="userUpdated")
+  #summary-wrapper
+    c-summary.widget-padding(
+      ref="summary",
+      v-bind:repos="users",
+      v-bind:error-messages="errorMessages",
+      v-bind:is-widget-mode="true"
+    )
+template(v-else)
+  .empty Widget does not support uploading of .zip file generated by RepoSense.
+</template>
+
+<script>
+import cSummary from './c-summary.vue';
+
+export default {
+  name: 'c-widget',
+  components: {
+    cSummary,
+  },
+  props: {
+    updateReportZip: {
+      type: Function,
+      required: true,
+    },
+    repos: {
+      type: Object,
+      required: true,
+    },
+    users: {
+      type: Array,
+      required: true,
+    },
+    userUpdated: {
+      type: Boolean,
+      required: true,
+    },
+    loadingOverlayOpacity: {
+      type: Number,
+      required: true,
+    },
+    tabType: {
+      type: String,
+      required: true,
+    },
+    creationDate: {
+      type: String,
+      required: true,
+    },
+    reportGenerationTime: {
+      type: String,
+      required: true,
+    },
+    errorMessages: {
+      type: Object,
+      required: true,
+    },
+  },
+};
+</script>
+
+<style lang="scss">
+@import '../styles/_colors.scss';
+@import '../styles/summary-chart.scss';
+</style>
